Everything a bit better plus updated tutorials
