+++ src/model_setup.py
@@ -1,22 +1,43 @@
-
+import copy
 import csv
-import pandas as pd
-import numpy as np
-import scanpy as sc
-from scipy.stats import norm, gamma
-import torch
-from torch import nn
-from torch.utils.data import Dataset, DataLoader
-import sparselinear as sl
+import math
+import os
+from importlib import resources
+
 import anndata
 import matplotlib.pyplot as plt
-import copy
-import os
-import random
-import math
-import scipy.stats
+import numpy as np
+import pandas as pd
+import scanpy as sc
+import scipy
+import sparselinear as sl
+import torch
+from scipy.stats import gamma, norm
+from torch import nn
+from torch.utils.data import DataLoader, Dataset
+from tqdm.auto import tqdm
 import ot
-from importlib import resources
+from output_paths import artifact_dir, artifact_path
+
+
+def _get_device(device=None):
+    if device is None:
+        return torch.device("cuda" if torch.cuda.is_available() else "cpu")
+    return torch.device(device)
+
+
+def _get_model_device(model, device=None):
+    if device is not None:
+        return _get_device(device)
+    try:
+        return next(model.parameters()).device
+    except StopIteration:
+        return _get_device()
+
+
+def _move_model_to_device(model, device=None):
+    resolved_device = _get_model_device(model, device=device)
+    return model.to(resolved_device), resolved_device
 
 
 def _module_path(*parts):
@@ -44,272 +65,169 @@ def _resolve_data_file(path):
     return path
 
 
+def _dense_matrix(adata):
+    return adata.X.toarray() if scipy.sparse.issparse(adata.X) else np.array(adata.X)
 
-def load_db(adata, file = 'human_lr_pair.tsv', sep='\t'):
+
+def _build_io_matrices(adata, lig_uni, rec_uni):
+    rec_keys = list(rec_uni.keys())
+    lig_keys = list(lig_uni.keys())
+    if scipy.sparse.issparse(adata.X):
+        input_recs = adata[:, rec_keys].X.toarray()
+        input_ligs = adata[:, lig_keys].X.toarray()
+    else:
+        input_recs = np.array(adata[:, rec_keys].X)
+        input_ligs = np.array(adata[:, lig_keys].X)
+    lig_averages = np.mean(input_ligs, axis=0)
+    inputs = np.concatenate((input_recs, np.repeat(lig_averages.reshape(1, -1), input_recs.shape[0], axis=0)), axis=1)
+    all_genes = _dense_matrix(adata)
+    return inputs, all_genes, input_recs, input_ligs
+
+
+def _loader_kwargs(device, shuffle=False, batch_size=256):
+    use_cuda = device.type == "cuda"
+    return {
+        "shuffle": shuffle,
+        "batch_size": batch_size,
+        "pin_memory": use_cuda,
+    }
+
+
+def _resolve_output_file(path, kind, default_name):
+    if path is None:
+        return artifact_path(kind, default_name)
+    if os.path.isabs(path):
+        parent = os.path.dirname(path)
+        if parent:
+            os.makedirs(parent, exist_ok=True)
+        return path
+    if os.path.dirname(path):
+        os.makedirs(os.path.dirname(path), exist_ok=True)
+        return path
+    if path == default_name:
+        return artifact_path(kind, default_name)
+    return path
+
+
+def load_db(adata, file="human_lr_pair.tsv", sep="\t"):
     print("Loading Database...")
     file = _resolve_data_file(file)
-    #human ligand universe
     lig_uni = {}
-    #human receptor universe
     rec_uni = {}
-    
-    with open(file, 'r') as csvfile:
+
+    with open(file, "r") as csvfile:
         datareader = csv.reader(csvfile, delimiter=sep)
-        i = 0
+        next(datareader, None)
+        genes = set(adata.var_names)
         for row in datareader:
             ligand = row[1]
             receptor = row[2]
-            lig_keys = lig_uni.keys()
-            rec_keys = rec_uni.keys()
-            #skip headers
-            if i == 0:
-                i = i + 1
-                continue
-            #ligand
-            if ligand in lig_keys and receptor in adata.var_names:
-                list1 = lig_uni[ligand]
-                list1.append(receptor)
-            elif ligand in adata.var_names and receptor in adata.var_names:
-                list1 = [receptor]
-                lig_uni[ligand] = list1
-            #receptor 
-            if receptor in rec_keys and ligand in adata.var_names:
-                list2 = rec_uni[receptor]
-                list2.append(ligand)
-            elif receptor in adata.var_names and ligand in adata.var_names:
-                list2 = [ligand]
-                rec_uni[receptor] = list2
-                
-    #List of LR pairs          
-    lr_pairs = []
-    for ligand in lig_uni.keys():
-        for receptor in lig_uni[ligand]:
-            lr_pairs.append(ligand+"_"+receptor)
+            if ligand in genes and receptor in genes:
+                lig_uni.setdefault(ligand, []).append(receptor)
+                rec_uni.setdefault(receptor, []).append(ligand)
+
+    lr_pairs = [f"{ligand}_{receptor}" for ligand, receptors in lig_uni.items() for receptor in receptors]
     print("Database Loaded!")
     return lig_uni, rec_uni, lr_pairs
 
 
-
-def delta_preprocesssing(adata, rec_uni, lig_uni):
-    #Preprocessing to obtain the expressions of every lig/rec and coordinates for each spot
-    lig_dist = {}
-    lig_coords = {}
-    rec_dist = {}
-    rec_coords = {}
-    genes = list(adata.var_names)
-
-    x = list(adata.obs['x_coord'])
-    y = list(adata.obs['y_coord'])
-    for lig in lig_uni:
-        j = genes.index(lig)
-        exp = adata.X.toarray()[:, j]
-        
-        d = {"exp":exp, "x":x,"y":y}
-        df2 = pd.DataFrame(data=d, index=cells)
-        df2 = df2[df2.exp > 0.0]
-        lig_dist[lig] = list(df2['exp'])
-        lig_coords[lig] = np.column_stack((list(df2['x']), np.array(list(df2['y']))))
-        
-
-    for rec in rec_uni:
-        j = genes.index(rec)
-        exp = adata.X.toarray()[:, j]
-        
-        d = {"exp":exp, "x":x,"y":y}
-        df2 = pd.DataFrame(data=d,index=cells)
-        df2 = df2[df2.exp > 0.0]
-        rec_dist[rec] = list(df2['exp'])
-        rec_coords[rec] = np.column_stack((list(df2['x']), np.array(list(df2['y']))))
-    
-    return lig_dist, rec_dist, lig_coords, rec_coords
-
-
-def calc_distance(pair, pos_l, pos_r, d_lig, d_rec, reg=1, reg_m=1, iters=100):
-    lig, rec = pair.split("_")
-    
-    #unbalanced OT - forward direction 
-    cost = ot.dist(pos_l, pos_r,metric='euclidean')
-    cost = cost +1e-8
-    distance_fwd = ot.sinkhorn_unbalanced2(d_lig,d_rec, cost/cost.max(), reg=reg, reg_m=reg_m, numItermax=iters)[0]
-    
-    #unbalanced OT - reverse direction
-    cost2 = ot.dist(pos_r, pos_l,metric='euclidean')
-    cost2 = cost2 +1e-8
-    distance_rev = ot.sinkhorn_unbalanced2(d_rec,d_lig, cost2/cost2.max(), reg=reg, reg_m=reg_m, numItermax=iters)[0]
-    
-    #Final Distance
-    gene_distance = (distance_fwd+distance_rev)/2
-    
-    return gene_distance
-
-
 def load_tf_db(species, adata, rec_uni):
-    if species == 'mouse':
+    if species == "mouse":
         file1 = _resolve_data_file("databases/TF_TG_mouse.csv")
         file2 = _resolve_data_file("databases/KEGG_mouse.csv")
         file3 = _resolve_data_file("databases/REACTOME_mouse.csv")
-    if species == 'human':
+    elif species == 'human':
         file1 = _resolve_data_file("databases/TF_TG_human.csv")
         file2 = _resolve_data_file("databases/KEGG_human.csv")
         file3 = _resolve_data_file("databases/REACTOME_human.csv")
-      
-    #TF_TG dataframe
-    df = pd.read_csv(file1)
-    df = df.drop(columns="Unnamed: 0")
-    
-    #KEGG dataframe
-    df2 = pd.read_csv(file2)
-    df2 = df2.drop(columns="Unnamed: 0")
-    
-    #REACTOME dataframe
-    df3 = pd.read_csv(file3)
-    df3 = df3.drop(columns="Unnamed: 0")
-    
-    #df2 cleanup
-    receptors = list(df2['receptor'])
-    tf = list(df2['tf'])
-    list1 = []
-    list2 = []
-    for i, rec in enumerate(receptors):
-        l = rec.split(",")
-        for rec2 in l:
-            list1.append(rec2)
-            list2.append(tf[i])
-            
-    df2 = pd.DataFrame({"Receptors":list1, "TF":list2})
-    
-    #df3 cleanup
-    receptors = list(df3['receptor'])
-    tf = list(df3['tf'])
-    list1 = []
-    list2 = []
-    for i, rec in enumerate(receptors):
-        l = rec.split(",")
-        for rec2 in l:
-            list1.append(rec2)
-            list2.append(tf[i])
-            
-    df3 = pd.DataFrame({"Receptors":list1, "TF":list2})
-    
+
+    df = pd.read_csv(file1).drop(columns="Unnamed: 0", errors="ignore")
+    df2 = pd.read_csv(file2).drop(columns="Unnamed: 0", errors="ignore")
+    df3 = pd.read_csv(file3).drop(columns="Unnamed: 0", errors="ignore")
+
+    def _explode_rec_tf(frame):
+        receptors = list(frame["receptor"])
+        tfs = list(frame["tf"])
+        list1 = []
+        list2 = []
+        for i, rec in enumerate(receptors):
+            for rec2 in rec.split(","):
+                list1.append(rec2)
+                list2.append(tfs[i])
+        return pd.DataFrame({"Receptors": list1, "TF": list2})
+
     #Rec to TF dataframe
-    rec_tf = pd.concat([df2, df3]).drop_duplicates()
-    
+    rec_tf = pd.concat([_explode_rec_tf(df2), _explode_rec_tf(df3)]).drop_duplicates()
+    genes = set(adata.var_names)
+
     #TF universe - TF to downstream genes
     tf_uni = {}
     for _, row in df.iterrows():
-        tf = row['TF']
-        if tf in adata.var_names:
-            gene = row['Target Gene']
-            if tf in tf_uni.keys():
-                gene_list = tf_uni[tf]
-            else:
-                gene_list = []
-            if gene in adata.var_names and gene not in gene_list:
-                gene_list.append(gene)
-                tf_uni[tf] = gene_list
-    
-    #Rec to TF universe    
+        tf = row["TF"]
+        gene = row["Target Gene"]
+        if tf in genes and gene in genes:
+            tf_uni.setdefault(tf, [])
+            if gene not in tf_uni[tf]:
+                tf_uni[tf].append(gene)
+
+    #Rec to TF universe  
     rec_tf_uni = {}
     for _, row in rec_tf.iterrows():
         rec = row["Receptors"]
         tf = row["TF"]
-        if rec in rec_tf_uni.keys():
-            tf_list = rec_tf_uni[rec]
-        else:
-            tf_list = []
-        if rec in rec_uni.keys() and tf in tf_uni.keys() and tf not in tf_list:
-            tf_list.append(tf)
-            rec_tf_uni[rec] = tf_list
+        if rec in rec_uni and tf in tf_uni:
+            rec_tf_uni.setdefault(rec, [])
+            if tf not in rec_tf_uni[rec]:
+                rec_tf_uni[rec].append(tf)
 
     return tf_uni, rec_tf_uni
 
 
-
 def create_masked_connections(adata, lig_uni, rec_uni, tf_uni, rec_tf_uni, lr_pairs):
-    #inputs
     num_rec = len(rec_uni)
     num_lig = len(lig_uni)
     num_inputs = num_rec + num_lig
-    l = list(rec_uni.keys())+list(lig_uni.keys())
-    #inputs = adata[:, l].X.toarray()
-    if scipy.sparse.issparse(adata.X):
-        input_recs = adata[:, list(rec_uni.keys())].X.toarray()
-        input_ligs = adata[:, list(lig_uni.keys())].X.toarray()
-    else:
-        input_recs = np.array(adata[:, list(rec_uni.keys())].X)
-        input_ligs = np.array(adata[:, list(lig_uni.keys())].X)
-    lig_averages = np.mean(input_ligs, axis=0)
-    reshaped_averages = lig_averages.reshape(1, -1)
-    repeated_averages = np.repeat(reshaped_averages, input_recs.shape[0], axis=0)
-    inputs = np.concatenate((input_recs, repeated_averages), axis=1)
+    input_names = list(rec_uni.keys()) + list(lig_uni.keys())
+    inputs, all_genes, _, _ = _build_io_matrices(adata, lig_uni, rec_uni)
 
-
-    #hidden layers
     num_lr_pairs = len(lr_pairs)
     num_tfs = len(tf_uni.keys())
-
-    #output
-    if scipy.sparse.issparse(adata.X):
-        all_genes = adata.X.toarray()
-    else:
-        all_genes = np.array(adata.X)
     num_outputs = all_genes.shape[1]
-    
-    #indexes
-    lr_index =dict(zip(lr_pairs, list(range(num_lr_pairs))))
-    input_index = dict(zip(l,list(range(num_inputs))))
-    tf_index = dict(zip(list(tf_uni.keys()),list(range(num_tfs))))
-    gene_index = dict(zip(list(adata.var_names),list(range(num_outputs))))
-    
-    #Layer 1 connections
-    mask = np.zeros((num_lr_pairs, num_inputs))
-    for pair, i in list(lr_index.items()):
-        lig, rec = pair.split("_")
-        #ligand index
-        j = input_index[lig]
-        #receptor index 
-        k = input_index[rec]
-        #set connection between lig and receptor to node in LR pair layer
-        mask[i][j] = 1
-        mask[i][k] = 1
-    mask = torch.nonzero(torch.tensor(mask)).T
-    
-    #Layer 2 Connections
-    mask2 = np.zeros((num_tfs,num_lr_pairs))
-    
-    for pair, j in list(lr_index.items()):
-        _, rec = pair.split("_")
-        if rec in rec_tf_uni.keys():
-            #get tfs for that receptor
-            tf_list = rec_tf_uni[rec]
-            for tf in tf_list:
-                i = tf_index[tf]
-                #set connection between LR pair with the receptor and TF
-                mask2[i][j] = 1
-        else:
-            vector = np.ones(mask2.shape[0])
-            mask2[:, j] = vector
-            
-    mask2 = mask2.T
-    
-    #Layer 3 Connections
-    mask3 = np.zeros((num_outputs,num_tfs))
-    for tf, j in list(tf_index.items()):
-        #get downstream genes for each TF
-        gene_list = tf_uni[tf]
-        for gene in gene_list:
-            i = gene_index[gene]
-            #set connections between TF and downstream gene
-            mask3[i][j] = 1
 
-    fc_genes = []
-    #for any downstream gene with no connection make them fully connected with previous layer
+    lr_index = dict(zip(lr_pairs, list(range(num_lr_pairs))))
+    input_index = dict(zip(input_names, list(range(num_inputs))))
+    tf_index = dict(zip(list(tf_uni.keys()), list(range(num_tfs))))
+    gene_index = dict(zip(list(adata.var_names), list(range(num_outputs))))
+
+    # Layer 1 Connections
+    mask = np.zeros((num_lr_pairs, num_inputs))
+    for pair, i in lr_index.items():
+        lig, rec = pair.split("_")
+        mask[i][input_index[lig]] = 1
+        mask[i][input_index[rec]] = 1
+    mask = torch.nonzero(torch.tensor(mask)).T
+
+    # Layer 2 Connections
+    mask2 = np.zeros((num_tfs, num_lr_pairs))
+    for pair, j in lr_index.items():
+        _, rec = pair.split("_")
+        if rec in rec_tf_uni:
+            for tf in rec_tf_uni[rec]:
+                mask2[tf_index[tf]][j] = 1
+        else:
+            mask2[:, j] = np.ones(mask2.shape[0])
+    mask2 = mask2.T
+
+    # Layer 3 Connections
+    mask3 = np.zeros((num_outputs, num_tfs))
+    for tf, j in tf_index.items():
+        for gene in tf_uni[tf]:
+            mask3[gene_index[gene]][j] = 1
+
+    # for any downstream gene with no connection make them fully connected with previous layer
     for i in range(num_outputs):
         if not np.any(mask3[i]):
-            vector = np.ones(len(mask3[i]))
-            mask3[i] = vector
-            fc_genes.append(adata.var_names[i])
-    
+            mask3[i] = np.ones(len(mask3[i]))
     mask3 = mask3.T
     return inputs, all_genes, num_inputs, num_lr_pairs, num_tfs, num_outputs, mask, mask2, mask3
 
@@ -323,26 +241,17 @@ def normalize(mat):
             if mmin == mmax:
                 val2 = 0
             else:
-                val2 = (val - mmin)/(mmax-mmin)
+                val2 = (val - mmin) / (mmax - mmin)
             row[j] = val2
         smat[i] = row
-    mat=smat.transpose()
+    mat = smat.transpose()
     return mat
 
 
-# Define custome autograd function for masked connection.
-
 class CustomizedLinearFunction(torch.autograd.Function):
-    """
-    autograd function which masks it's weights by 'mask'.
-    """
-
-    # Note that both forward and backward are @staticmethods
     @staticmethod
-    # bias, mask is an optional argument
     def forward(ctx, input, weight, bias=None, mask=None):
         if mask is not None:
-            # change weight to 0 where mask == 0
             weight = weight * mask
         output = input.mm(weight.t())
         if bias is not None:
@@ -350,29 +259,17 @@ class CustomizedLinearFunction(torch.autograd.Function):
         ctx.save_for_backward(input, weight, bias, mask)
         return output
 
-    # This function has only a single output, so it gets only one gradient
     @staticmethod
     def backward(ctx, grad_output):
-        # This is a pattern that is very convenient - at the top of backward
-        # unpack saved_tensors and initialize all gradients w.r.t. inputs to
-        # None. Thanks to the fact that additional trailing Nones are
-        # ignored, the return statement is simple even when the function has
-        # optional inputs.
         input, weight, bias, mask = ctx.saved_tensors
         grad_input = grad_weight = grad_bias = grad_mask = None
 
-        # These needs_input_grad checks are optional and there only to
-        # improve efficiency. If you want to make your code simpler, you can
-        # skip them. Returning gradients for inputs that don't require it is
-        # not an error.
         if ctx.needs_input_grad[0]:
             grad_input = grad_output.mm(weight)
         if ctx.needs_input_grad[1]:
             grad_weight = grad_output.t().mm(input)
             if mask is not None:
-                # change grad_weight to 0 where mask == 0
                 grad_weight = grad_weight * mask
-        #if bias is not None and ctx.needs_input_grad[2]:
         if ctx.needs_input_grad[2]:
             grad_bias = grad_output.sum(0).squeeze(0)
 
@@ -381,19 +278,7 @@ class CustomizedLinearFunction(torch.autograd.Function):
 
 class CustomizedLinear(nn.Module):
     def __init__(self, mask, bias=True):
-        """
-        extended torch.nn module which mask connection.
-
-        Argumens
-        ------------------
-        mask [torch.tensor]:
-            the shape is (n_input_feature, n_output_feature).
-            the elements are 0 or 1 which declare un-connected or
-            connected.
-        bias [bool]:
-            flg of bias.
-        """
-        super(CustomizedLinear, self).__init__()
+        super().__init__()
         self.input_features = mask.shape[0]
         self.output_features = mask.shape[1]
         if isinstance(mask, torch.Tensor):
@@ -402,50 +287,31 @@ class CustomizedLinear(nn.Module):
             self.mask = torch.tensor(mask, dtype=torch.float).t()
 
         self.mask = nn.Parameter(self.mask, requires_grad=False)
-
-        # nn.Parameter is a special kind of Tensor, that will get
-        # automatically registered as Module's parameter once it's assigned
-        # as an attribute. Parameters and buffers need to be registered, or
-        # they won't appear in .parameters() (doesn't apply to buffers), and
-        # won't be converted when e.g. .cuda() is called. You can use
-        # .register_buffer() to register buffers.
-        # nn.Parameters require gradients by default.
         self.weight = nn.Parameter(torch.Tensor(self.output_features, self.input_features))
 
         if bias:
             self.bias = nn.Parameter(torch.Tensor(self.output_features))
         else:
-            # You should always register all possible parameters, but the
-            # optional ones can be None if you want.
-            self.register_parameter('bias', None)
+            self.register_parameter("bias", None)
         self.reset_parameters()
-
-        # mask weight
         self.weight.data = self.weight.data * self.mask
 
     def reset_parameters(self):
-        stdv = 1. / math.sqrt(self.weight.size(1))
+        stdv = 1.0 / math.sqrt(self.weight.size(1))
         self.weight.data.uniform_(-stdv, stdv)
         if self.bias is not None:
             self.bias.data.uniform_(-stdv, stdv)
 
-
     def forward(self, input):
-        # See the autograd section for explanation of what happens here.
         return CustomizedLinearFunction.apply(input, self.weight, self.bias, self.mask)
 
     def extra_repr(self):
-        # (Optional)Set the extra information about this module. You can test
-        # it by printing an object of this class.
-        return 'input_features={}, output_features={}, bias={}'.format(
-            self.input_features, self.output_features, self.bias is not None
-        )
+        return f"input_features={self.input_features}, output_features={self.output_features}, bias={self.bias is not None}"
 
-    
- 
+
 class net(nn.Module):
-    def __init__(self,num_inputs, num_lr_pairs, num_tfs, output_size, mask1, mask2, mask3):
-        super(net,self).__init__()
+    def __init__(self, num_inputs, num_lr_pairs, num_tfs, output_size, mask1, mask2, mask3):
+        super().__init__()
         self.layers = nn.Sequential(
             sl.SparseLinear(in_features=num_inputs, out_features=num_lr_pairs, connectivity=mask1),
             nn.BatchNorm1d(num_lr_pairs),
@@ -453,149 +319,45 @@ class net(nn.Module):
             CustomizedLinear(torch.tensor(mask2), bias=None),
             nn.BatchNorm1d(num_tfs),
             CustomizedLinear(torch.tensor(mask3), bias=None),
-            nn.ReLU()
+            nn.ReLU(),
         )
-        
-    def forward(self,x):
+
+    def forward(self, x):
         output = self.layers(x)
         return output
-    
-    
-class ExpDataset():
-    def __init__(self,x,y):
-        self.x = torch.tensor(x,dtype=torch.float32)
-        self.y = torch.tensor(y,dtype=torch.float32)
+
+
+class ExpDataset(Dataset):
+    def __init__(self, x, y, device=None):
+        # Keep dataset tensors on CPU so DataLoader pin_memory can stage
+        # batches efficiently before the training loop moves them to CUDA.
+        self.x = torch.as_tensor(x, dtype=torch.float32).cpu()
+        self.y = torch.as_tensor(y, dtype=torch.float32).cpu()
         self.length = self.x.shape[0]
-    def __getitem__(self,idx):
-        return self.x[idx],self.y[idx]
+
+    def __getitem__(self, idx):
+        return self.x[idx], self.y[idx]
+
     def __len__(self):
         return self.length
 
 
-
-    
-def train(adata, lig_uni, rec_uni, tf_uni, rec_tf_uni, lr_pairs, path, epochs=50):
-    print("Setting up model")
-    o = create_masked_connections(adata, lig_uni, rec_uni, tf_uni, rec_tf_uni, lr_pairs)
-    inputs, all_genes, num_inputs, num_lr_pairs, num_tfs, num_outputs, mask, mask2, mask3 = o
-    dataset = ExpDataset(inputs, all_genes)
-    dataloader = DataLoader(dataset=dataset,shuffle=False,batch_size=256)
-    model = net(num_inputs, num_lr_pairs, num_tfs, num_outputs, mask, mask2, mask3)
-    criterion = nn.MSELoss()
-    optimizer = torch.optim.SGD(model.parameters(),lr=0.1)
-    
-    #training
-    print("Training model...")
-    preds = []
-    costval = []
-    for j in range(epochs):
-        print("epoch: "+str(j))
-        for i,(x_train,y_train) in enumerate(dataloader):
-            optimizer.zero_grad()
-            #prediction
-            y_pred = model(x_train)
-            #calculating loss
-            cost = criterion(y_pred,y_train)
-            #backprop
-            cost.backward()
-            optimizer.step()
-        print(cost)
-        costval.append(cost)
-    torch.save(model, path)
-    print("Training complete!")
-    return inputs, all_genes
-    
-    
-def load_model(path):
-    model = torch.load(path, weights_only=False)
-    return model
-
-
-def perform_iteration(model, mat, C, criterion, batch = 256):
-    dataset2 = ExpDataset(mat,C)
-    dataloader2 = DataLoader(dataset=dataset2,shuffle=False, batch_size=batch)
-    for i,(x_train,y_train) in enumerate(dataloader2):
-        y_pred = model(x_train)
-    cost = criterion(y_pred,y_train)
-    return y_pred, cost
-        
-
-
-def permutate_receptors(model, start, num, ocost, mat, C, criterion, loss=[], perc=50):
-    zeros = np.zeros(len(mat))
-    for i in range(start, num):
-        print(i)
-        mat2 = copy.deepcopy(mat)
-        mat2 = mat2.transpose()
-        tmp = mat2[i]
-        random.shuffle(tmp)
-        new_arr_no_0 = tmp[tmp!=0.0]
-        if perc == 100:
-            tmp = zeros
-        elif len(new_arr_no_0) != 0:
-            scale = np.percentile(new_arr_no_0, 100-perc)
-            if scale < 1:
-                tmp = tmp*scale
-            else:
-                tmp = tmp/scale
-        mat2[i] = tmp
-        mat2 = mat2.transpose()
-        ypred, cost = perform_iteration(model, mat2, C, criterion, batch = mat2.shape[0])
-        dcost = abs(cost - ocost)
-        loss.append(dcost)
-    return loss
-
-
-def save_conversion_rate(conversion_rates, file='conversion_rates.txt'):
-    with open(file, 'w') as fp:
+def save_conversion_rate(conversion_rates, file="conversion_rates.txt"):
+    file = _resolve_output_file(file, "conversion_rates", "conversion_rates.txt")
+    with open(file, "w") as fp:
         for values in conversion_rates:
-            # write each item on a new line
-            fp.write("%s\n" % values)
-    print('Done')
-    
-    
-def load_conversion_rate(file='conversion_rates.txt'):
-    conversion_rates = []
-    with open(file, 'r') as fp:
-        for line in fp:
-            # remove linebreak from a current name
-            # linebreak is the last character of each line
-            x = line[:-1]
+            fp.write(f"{values}\n")
+    print("Done")
 
-            # add current item to the list
-            conversion_rates.append(float(x))
+
+def load_conversion_rate(file="conversion_rates.txt"):
+    file = _resolve_output_file(file, "conversion_rates", "conversion_rates.txt")
+    conversion_rates = []
+    with open(file, "r") as fp:
+        for line in fp:
+            conversion_rates.append(float(line[:-1]))
     print("Conversion rates loaded")
     return conversion_rates
-        
-
-def feature_selection(model, mat, C, rec_uni, start=0, perc=50):
-    print("Performing feature selection to obtain conversion rates...")
-    #one iteration of model
-    criterion = nn.MSELoss()
-    dataset2 = ExpDataset(mat,C)
-    dataloader2 = DataLoader(dataset=dataset2,shuffle=False,batch_size=407)
-    for i,(x_train,y_train) in enumerate(dataloader2):
-        y_pred = model(x_train)
-    cost = criterion(y_pred,y_train)
-    
-    losses = permutate_receptors(
-        model, start=0, num=len(rec_uni), mat=mat, C=C, ocost=cost, loss=[], criterion=criterion, perc=perc)
-    mean = torch.mean(torch.stack(losses))
-    lmin = min(losses)
-    lmax = max(losses)
-    l2 = [val/mean for val in losses]
-    conversion_rates = []
-    for val in l2:
-        if val > 1:
-            conversion_rates.append(1)
-        elif val < 0.4:
-            conversion_rates.append(0.4)
-        else: 
-            conversion_rates.append((val.item()))
-
-    print("Complete")
-    return conversion_rates
-
 
 
 def add_rates(conversion_rates, rec_uni):
@@ -603,113 +365,298 @@ def add_rates(conversion_rates, rec_uni):
     return rates
 
 
-
-def get_target_genes(receptors, N, model, mat, C, rec_uni, adata, perc=50, threshold=50):
-    dict2 = {}
-    y_pred, _ = perform_iteration(model, mat, C, nn.MSELoss(), batch = mat.shape[0])
-    for rec in receptors:
-        print("Getting Target Genes for receptor: " +rec)
-        zeros = np.zeros(N)
-        mat2 = copy.deepcopy(mat)
-        mat2 = mat2.transpose()
-        matched = False
-        for i, rec2 in enumerate(rec_uni.keys()):
-            if rec2 == rec:
-                tmp = mat2[i]
-                if perc != 100:
-                    random.shuffle(tmp)
-                    new_arr_no_0 = tmp[tmp!=0.0]
-                    if len(new_arr_no_0) != 0:
-                        scale = np.percentile(new_arr_no_0, 100-perc)
-                        if scale < 1:
-                            tmp = tmp*scale
-                        else:
-                            tmp = tmp/scale
-                else:
-                    tmp = zeros
-                mat2[i] = tmp
-                matched = True
-                break
-        if not matched:
-            raise ValueError(f"Receptor {rec} not found in rec_uni")
-        mat2 = mat2.transpose()        
-        """
-                mat2[i] = zeros
-        mat2 = mat2.transpose()
-        mat2 = standardize(mat2)
-        """
-        y_pred2, _ = perform_iteration(model, mat2, C, nn.MSELoss(), batch = mat2.shape[0])
-        y = abs(y_pred-y_pred2)
-        y = y.cpu().detach().numpy()
-        zz = y.mean(axis=0)
-        
-        ind = np.argpartition(zz, -threshold)[-threshold:]
-
-        top50 = list(zz[ind])
-        new=[]
-        for i, gene in enumerate(adata.var_names):
-            if i in ind:
-                new.append(gene)
-        #new
-        dict2[rec] = (top50, new)
-    return dict2
-    
+def load_model(path="model.pt", device=None):
+    path = _resolve_output_file(path, "models", "model.pt")
+    resolved_device = _get_device(device)
+    model = torch.load(path, map_location=resolved_device, weights_only=False)
+    return model.to(resolved_device)
 
 
-def plot_results(dict1, adata, path='figures'):
-    isExist = os.path.exists(path)
-    if not isExist:
-       # Create a new directory because it does not exist
-       os.makedirs(path)
+def plot_results(dict1, adata, path="figures"):
+    path = artifact_dir("perturbation") if path == "figures" else path
+    os.makedirs(path, exist_ok=True)
     for rec in dict1.keys():
-        print("Plotting results for receptor: "+rec)
-        # Barplot
+        print("Plotting results for receptor: " + rec)
+
+        # Barplot 
         genes = dict1[rec][1]
-        values = dict1[rec][0]
-        values = [x*100 for x in values]
-        genes2 = [y for _,y in sorted(zip(values,genes), reverse=True)]
+        values = [x * 100 for x in dict1[rec][0]]
+        genes2 = [y for _, y in sorted(zip(values, genes), reverse=True)]
         values.sort(reverse=True)
-        plt.figure(figsize=(10,7))
+        plt.figure(figsize=(10, 7))
         plt.xticks(rotation=90)
         plt.xlabel("Target Genes")
         plt.ylabel("Percent Change in Gene Expression")
-        plt.ylim((min(values),max(values)+0.05))
+        plt.ylim((min(values), max(values) + 0.05))
         plt.bar(genes2, values)
-        plt.savefig(path+'/'+rec+'_targets.pdf')
-        
-        
-        # Matrix plot
-        sc.pl.matrixplot(adata, genes2, groupby='cell_type', cmap='viridis',standard_scale='var',save=rec+'_matrix.pdf')
-    print("plots saved to directory /"+path)
+        plt.savefig(path + "/" + rec + "_targets.pdf")
+        plt.close()
+
+        # Matrix Plot
+        prev_figdir = sc.settings.figdir
+        try:
+            sc.settings.figdir = path
+            sc.pl.matrixplot(
+                adata,
+                genes2,
+                groupby="cell_type",
+                cmap="viridis",
+                standard_scale="var",
+                save=rec + "_matrix.pdf",
+            )
+        finally:
+            sc.settings.figdir = prev_figdir
+    print("plots saved to directory /" + path)
 
 
+def train(adata, lig_uni, rec_uni, tf_uni, rec_tf_uni, lr_pairs, path="model.pt", epochs=50, lr=0.1, batch=256, device=None):
+    print("Setting up model")
+    path = _resolve_output_file(path, "models", "model.pt")
+    device = _get_device(device)
+    model_components = create_masked_connections(adata, lig_uni, rec_uni, tf_uni, rec_tf_uni, lr_pairs)
+    inputs, all_genes, num_inputs, num_lr_pairs, num_tfs, num_outputs, mask, mask2, mask3 = model_components
+    dataset = ExpDataset(inputs, all_genes, device=device)
+    dataloader = DataLoader(dataset=dataset, **_loader_kwargs(device, shuffle=False, batch_size=batch))
+    model = net(num_inputs, num_lr_pairs, num_tfs, num_outputs, mask, mask2, mask3).to(device)
+    criterion = nn.MSELoss()
+    optimizer = torch.optim.SGD(model.parameters(), lr=lr)
 
-def block_receptors(adata, receptors, rec_uni, lig_uni, net, perc=50, threshold=50):
-    print("Blocked Receptor Analysis")
-    model = load_model(net)
+    print("Training model...")
+    epoch_losses = []
+    epoch_iter = tqdm(range(epochs), desc="Training epochs", unit="epoch")
+    for _ in epoch_iter:
+        last_cost = None
+        for x_train, y_train in dataloader:
+            x_train = x_train.to(device, non_blocking=device.type == "cuda")
+            y_train = y_train.to(device, non_blocking=device.type == "cuda")
+            optimizer.zero_grad()
+            y_pred = model(x_train)
+            cost = criterion(y_pred, y_train)
+            cost.backward()
+            optimizer.step()
+            last_cost = cost
+        if last_cost is not None:
+            epoch_iter.set_postfix(loss=f"{last_cost.item():.6f}")
+            epoch_losses.append(last_cost.detach().cpu())
+    model = model.to("cpu")
+    torch.save(model, path)
+    print("Training complete!")
+    return inputs, all_genes
+
+
+def perform_iteration(model, mat, C, criterion, batch=256, device=None, return_predictions=True):
+    model, device = _move_model_to_device(model, device=device)
+    dataset2 = ExpDataset(mat, C, device=device)
+    dataloader2 = DataLoader(dataset=dataset2, **_loader_kwargs(device, shuffle=False, batch_size=batch))
+    was_training = model.training
     model.eval()
-    #get NN inputs
-    if scipy.sparse.issparse(adata.X):
-        input_recs = adata[:, list(rec_uni.keys())].X.toarray()
-        input_ligs = adata[:, list(lig_uni.keys())].X.toarray()
-    else:
-        input_recs = np.array(adata[:, list(rec_uni.keys())].X)
-        input_ligs = np.array(adata[:, list(lig_uni.keys())].X)
-    lig_averages = np.mean(input_ligs, axis=0)
-    reshaped_averages = lig_averages.reshape(1, -1)
-    repeated_averages = np.repeat(reshaped_averages, input_recs.shape[0], axis=0)
-    inputs = np.concatenate((input_recs, repeated_averages), axis=1)
+    predictions = [] if return_predictions else None
+    total_loss = torch.zeros((), device=device)
+    total_examples = 0
+    with torch.inference_mode():
+        for x_train, y_train in dataloader2:
+            x_train = x_train.to(device, non_blocking=device.type == "cuda")
+            y_train = y_train.to(device, non_blocking=device.type == "cuda")
+            batch_pred = model(x_train)
+            batch_cost = criterion(batch_pred, y_train)
+            batch_size = x_train.shape[0]
+            if return_predictions:
+                predictions.append(batch_pred)
+            total_loss += batch_cost * batch_size
+            total_examples += batch_size
 
-    #output
-    if scipy.sparse.issparse(adata.X):
-        all_genes = adata.X.toarray()
-    else:
-        all_genes = np.array(adata.X)
-    
+    if total_examples == 0:
+        raise ValueError("perform_iteration received no data to iterate over")
+
+    y_pred = torch.cat(predictions, dim=0) if return_predictions else None
+    cost = total_loss / total_examples
+    if was_training:
+        model.train()
+    return y_pred, cost
+
+
+def _shuffle_array(values, rng):
+    shuffled = np.array(values, copy=True)
+    rng.shuffle(shuffled)
+    return shuffled
+
+
+def _apply_receptor_perturbation(values, perc, zeros, rng):
+    perturbed_values = _shuffle_array(values, rng)
+    nonzero_values = perturbed_values[perturbed_values != 0.0]
+    if perc == 100:
+        return zeros
+    if len(nonzero_values) != 0:
+        scale = np.percentile(nonzero_values, 100 - perc)
+        perturbed_values = perturbed_values * scale if scale < 1 else perturbed_values / scale
+    return perturbed_values
+
+
+def permutate_receptors(
+    model,
+    start,
+    num,
+    ocost,
+    mat,
+    C,
+    criterion,
+    loss=None,
+    perc=100,
+    n_shuffles=10,
+    batch=256,
+    seed=None,
+    device=None,
+):
+    if loss is None:
+        loss = []
+    model, device = _move_model_to_device(model, device=device)
+    if n_shuffles < 1:
+        raise ValueError("n_shuffles must be at least 1")
+
+    zeroed_values = np.zeros(len(mat))
+    base_rng = np.random.default_rng(seed)
+    for receptor_idx in tqdm(range(start, num), desc="Permuting receptors", unit="receptor"):
+        receptor_losses = []
+        for _ in range(n_shuffles):
+            perturbed_matrix = copy.deepcopy(mat).transpose()
+            perturbed_values = _apply_receptor_perturbation(perturbed_matrix[receptor_idx], perc, zeroed_values, base_rng)
+            perturbed_matrix[receptor_idx] = perturbed_values
+            perturbed_matrix = perturbed_matrix.transpose()
+            _, perturbed_cost = perform_iteration(
+                model,
+                perturbed_matrix,
+                C,
+                criterion,
+                batch=batch,
+                device=device,
+                return_predictions=False,
+            )
+            receptor_losses.append(abs(perturbed_cost - ocost))
+        loss.append(torch.stack(receptor_losses).mean())
+    return loss
+
+
+def feature_selection(model, mat, C, rec_uni, start=0, perc=100, n_shuffles=10, batch=256, seed=None, device=None):
+    print("Performing feature selection to obtain conversion rates...")
+    model, device = _move_model_to_device(model, device=device)
+    criterion = nn.MSELoss()
+    _, cost = perform_iteration(
+        model, mat, C, criterion, batch=batch, device=device, return_predictions=False
+    )
+
+    losses = permutate_receptors(
+        model,
+        start=start,
+        num=len(rec_uni),
+        mat=mat,
+        C=C,
+        ocost=cost,
+        loss=[],
+        criterion=criterion,
+        perc=perc,
+        n_shuffles=n_shuffles,
+        batch=batch,
+        seed=seed,
+        device=device,
+    )
+    mean_loss = torch.mean(torch.stack(losses))
+    normalized_losses = [loss_value / mean_loss for loss_value in losses]
+    conversion_rates = []
+    for normalized_loss in normalized_losses:
+        if normalized_loss > 1:
+            conversion_rates.append(1)
+        elif normalized_loss < 0.4:
+            conversion_rates.append(0.4)
+        else:
+            conversion_rates.append(normalized_loss.item())
+
+    print("Complete")
+    return conversion_rates
+
+
+def get_target_genes(
+    receptors,
+    N,
+    model,
+    mat,
+    C,
+    rec_uni,
+    adata,
+    perc=100,
+    threshold=50,
+    n_shuffles=10,
+    batch=256,
+    seed=None,
+    device=None,
+):
+    target_gene_results = {}
+    model, device = _move_model_to_device(model, device=device)
+    if n_shuffles < 1:
+        raise ValueError("n_shuffles must be at least 1")
+
+    baseline_predictions, _ = perform_iteration(model, mat, C, nn.MSELoss(), batch=batch, device=device)
+    receptor_names = list(rec_uni.keys())
+    base_rng = np.random.default_rng(seed)
+    for rec in tqdm(receptors, desc="Getting target genes", unit="receptor"):
+        zeroed_values = np.zeros(N)
+        shuffled_predictions = []
+        for _ in range(n_shuffles):
+            perturbed_matrix = copy.deepcopy(mat).transpose()
+            for receptor_idx, receptor_name in enumerate(receptor_names):
+                if receptor_name != rec:
+                    continue
+                perturbed_values = _apply_receptor_perturbation(
+                    perturbed_matrix[receptor_idx], perc, zeroed_values, base_rng
+                )
+                perturbed_matrix[receptor_idx] = perturbed_values
+                break
+            perturbed_matrix = perturbed_matrix.transpose()
+            shuffled_prediction, _ = perform_iteration(
+                model, perturbed_matrix, C, nn.MSELoss(), batch=batch, device=device
+            )
+            shuffled_predictions.append(shuffled_prediction)
+        mean_perturbed_prediction = torch.stack(shuffled_predictions).mean(dim=0)
+        prediction_diff = abs(baseline_predictions - mean_perturbed_prediction).detach().cpu().numpy()
+        mean_gene_diff = prediction_diff.mean(axis=0)
+        top_gene_indices = np.argpartition(mean_gene_diff, -threshold)[-threshold:]
+        top_gene_scores = list(mean_gene_diff[top_gene_indices])
+        top_gene_names = [gene for gene_idx, gene in enumerate(adata.var_names) if gene_idx in top_gene_indices]
+        target_gene_results[rec] = (top_gene_scores, top_gene_names)
+    return target_gene_results
+
+
+def block_receptors(
+    adata,
+    receptors,
+    rec_uni,
+    lig_uni,
+    net,
+    perc=100,
+    threshold=50,
+    n_shuffles=10,
+    batch=256,
+    seed=None,
+    device=None,
+):
+    print("Blocked Receptor Analysis")
+    device = _get_device(device)
+    model = load_model(net, device=device)
+    model.eval()
+    inputs, all_genes, _, _ = _build_io_matrices(adata, lig_uni, rec_uni)
     print("Get original")
-    dict1 = get_target_genes(receptors, len(adata.obs), model, inputs, all_genes, rec_uni, adata, perc, threshold)
-    plot_results(dict1, adata)
-    
-    
-    
-    
+    target_gene_results = get_target_genes(
+        receptors,
+        len(adata.obs),
+        model,
+        inputs,
+        all_genes,
+        rec_uni,
+        adata,
+        perc=perc,
+        threshold=threshold,
+        n_shuffles=n_shuffles,
+        batch=batch,
+        seed=seed,
+        device=device,
+    )
+    plot_results(target_gene_results, adata)
